Trying to figure out why this behaves the way it does.
The main point is that I rewrote add from the simplistic version that doesn't use rbp.
Now in doing that I forgot to do popq %rbp before returning.

Usually, an error like that causes the program to segfault because ret takes the value
that's on top of the stack and makes the program go execute what's at that address.

The problem is that the program continues execution somewhere that is OK.

Other point, if in main, I pushq %rbp and do movq %rsp, %rbp, then that changes what's in
old rbp when we're in the function add, old ebp in that case is the value that %rsp had when
we entered in main.  In that case, executing the program causes the OS to terminate us
with the message Illegal instruction.  Indeed, old ebp contains, in this case a memory address
that which might make no sense.

The behavior that I see is that the program seems (and I do say seems) to resume execution
at the jne test_failed line (because if I change it, the return value of the program
changes.)

So the question is, if I don't touch %rbp in main, then what address does the old ebp in add
contain.  Because that is where we jump to.